finaly

+++ src/application/banner/components/banner-list.vue
@@ -133,30 +133,43 @@ export default {
                 style: {
                   color: 'red'
                 },
+                domProps: {
+                  innerHTML: '删除'
+                },
                 on: {
                   click: () => {
-                    axios.delete(`banner/${params.row._id}`).then(response => {
-                      // console.log(response.data)
-                      if (response.data.code === 200) {
-                        this.getOrSearchBannerList(1)
-                        this.$Message.success({
-                          content: response.data.msg,
-                          duration: 5,
-                          closable: true
+                    this.$Modal.confirm({
+                      title: '提示',
+                      content: '<p style="font-size: 14px">此操作将会永久删除该条记录，是否继续？</p>',
+                      closable: true,
+                      onOk: () => {
+                        axios.delete(`banner/${params.row._id}`).then(response => {
+                          // console.log(response.data)
+                          if (response.data.code === 200) {
+                            this.getOrSearchBannerList(1)
+                            this.$Message.success({
+                              content: response.data.msg,
+                              duration: 5,
+                              closable: true
+                            })
+                          } else {
+                            this.$Message.error({
+                              content: response.data.msg,
+                              duration: 5,
+                              closable: true
+                            })
+                          }
+                        }).catch(err => {
+                          console.log(err)
                         })
-                      } else {
-                        this.$Message.error({
-                          content: response.data.msg,
-                          duration: 5,
-                          closable: true
-                        })
+                      },
+                      onCancel: () => {
+                        console.log('取消了')
                       }
-                    }).catch(err => {
-                      console.log(err)
                     })
                   }
                 }
-              }, '删除')
+              })
             ])
           }
         }
